changed lightSensor.check to return an int

+++ Awesome/Awesome.h
@@ -68,7 +68,7 @@ class redGreenBlueLED {
 class LightSensor {
   public:
     void setup(int pin);
-    double check();
+    int check();
   private:
     int _pin;
 };
